refactor(translator): extract response.completed synthesis, fix ratchets
Fixes the file-size and complexity/cognitive-complexity ratchet
regressions the dedup-guard commit (6bbff5ea) introduced, so the PR's
own validation block (typecheck, eslint, file-size, complexity,
cognitive-complexity, changelog-integrity, test-discovery) is fully
green, not just its own tests:

- Extract the response.completed batched-tool-call synthesis body into
  a new leaf module
  open-sse/translator/response/openai-responses/synthesizeCompletedToolCalls.ts,
  mirroring this file's own established eventEmitter.ts/toolSchemas.ts/
  pureHelpers.ts extraction pattern, further split internally
  (buildToolCallChunks/buildFinalChunk/resolveArgsStr/baseChunk) to
  keep synthesizeCompletedToolCalls() itself under the complexity/
  cognitive-complexity/max-lines-per-function thresholds.
- computeFinishReason moves alongside it (not into the sibling
  pureHelpers.ts) because it takes stream `state` — pureHelpers.ts is
  guarded by tests/unit/response-openai-responses-purehelpers-split.test.ts
  to have NO state coupling at all.
- DRY the withAssistantRoleOnFirstDelta array/single-result branches
  into a shared setAssistantRoleIfEligible(state, delta) helper — the
  array branch alone pushed this function's cyclomatic complexity to
  16 (over the 15 threshold).
- Split the 5 new response.completed tests out of
  tests/unit/translator-resp-openai-responses.test.ts (which would
  have exceeded the 800-line test-file-size cap) into a new sibling
  tests/unit/translator-resp-openai-responses-completed-synthesis.test.ts.
- Bump the frozen open-sse/translator/response/openai-responses.ts
  file-size baseline for the small irreducible remainder (dedup-guard
  state tracking + call-site wiring) with a justification entry in
  config/quality/file-size-baseline.json.

Independently re-verified red-first (temporarily reintroducing the
pre-guard filter reproduces exactly one failure — the dedup test — no
collateral damage) and confirmed check:complexity-ratchets is back to
exactly baseline (2058/890) and check:file-size is fully green.
