+++ config/quality/file-size-baseline.json
@@ -8,6 +8,7 @@
   "_rebaseline_2026_07_03_v3844_ipfilter_release_green": "testFrozen bumps: models-catalog-route 1507->1600, perplexity-web 959->999, route-edge-coverage 1234->1241 (last is my #5975 comment +7). v3.8.44 cycle drift measured on release tip 32e4c906e during the #6131/#5975 release-green rebaseline. Inherited from the merge burst (Quality Ratchet does not run on PR->release fast-gates). route-edge-coverage +7 is my #5975 test comment; the rest is parallel-session drift. Tighten via --update next cycle.",
   "_rebaseline_2026_07_03_v3844_residual_release_green": "Residual file-size drift on tip 716041223: providerLimits.ts 955->982 + accountFallback.ts 1790->1864 (production god-files grown by parallel-session merges e.g. #6128; ideally DECOMPOSE not rebaseline, tracked as debt) + sse-auth.test.ts 1553->1600. None mine.",
   "_comment": "Catraca de tamanho (check-file-size.mjs). frozen so pode encolher; arquivos novos <= cap. --update ratcheta.",
+  "_rebaseline_2026_07_18_7613_completed_tool_calls_dedup": "PR #7613 (ekinnee) own growth: open-sse/translator/response/openai-responses.ts 1088->1112 (+24 = the toolCallIdsSeen dedup-guard tracking added at both the output_item.added and output_item.done handlers, so response.completed can skip re-synthesizing function_call items already streamed incrementally, plus the array-handling branch in withAssistantRoleOnFirstDelta for synthesized-chunk arrays (DRY'd into a shared setAssistantRoleIfEligible(state, delta) helper to keep its cyclomatic complexity under 15 — the array branch alone pushed it to 16), plus the new multi-line call-site wiring to synthesizeCompletedToolCalls()/computeFinishReason import). The ~130-line response.completed batched-tool-call synthesis body itself was EXTRACTED into a new leaf module open-sse/translator/response/openai-responses/synthesizeCompletedToolCalls.ts (164 LOC, well under cap; further split internally into buildToolCallChunks/buildFinalChunk/resolveArgsStr/baseChunk helpers to keep the exported synthesizeCompletedToolCalls() itself under the complexity/cognitive-complexity/max-lines-per-function thresholds — the initial single-function extraction still tripped all three), mirroring this file's own established eventEmitter.ts/toolSchemas.ts/pureHelpers.ts extraction pattern rather than growing the frozen god-file further. computeFinishReason (also extracted) lives in this new module rather than the sibling pureHelpers.ts because it takes stream `state` — pureHelpers.ts is guarded by tests/unit/response-openai-responses-purehelpers-split.test.ts to have NO state coupling at all (pureHelpers.ts itself is untouched, byte-identical, still 199 LOC). check:complexity-ratchets is back to exactly baseline (2058/890) after these two extractions — only file-size needed this bump. Covered by tests/unit/translator-resp-openai-responses.test.ts plus the new tests/unit/translator-resp-openai-responses-completed-synthesis.test.ts (also split out for the 800-line test-file-size cap), including a dedup regression test proving no duplicate tool_calls when a provider streams incrementally AND response.completed echoes the same function_call items in output[].",
   "_rebaseline_2026_07_09_pr6647_winget_claude_detect": "PR #6647 (enjoyer-hub, /implement-prs sync): cliRuntime.ts 1100->1110 (split('\\n').length metric; +10, was already exactly at the 1100 frozen cap). Adds the WinGet-installed Claude Code fallback path (%LOCALAPPDATA%\\Microsoft\\WinGet\\Packages\\Anthropic.ClaudeCode_Microsoft.Winget.Source_8wekyb3d8bbwe\\claude.exe) to getKnownToolPaths() alongside the two sibling Claude Code paths, so WinGet installs are auto-detected without CLI_CLAUDE_BIN. The package folder name (62 chars) forces Prettier's 100-char width to break the path.join call across the full 10-line multi-arg form used elsewhere in this same function for long paths; irreducible without changing the shared getKnownToolPaths() structure. Covered by the PR's own regression test (tests/unit/cli-runtime-detection.test.ts, win32-gated).",
   "_rebaseline_2026_07_03_review_prs_release_green": "Release-green unblock (2026-07-03, /review-prs): the quality.yml fast-gates job was base-red for EVERY PR->release from growth inherited via already-merged PRs on the release tip — no offending PR branch left to fix in-place. Prod frozen raised: ApiManagerPageClient.tsx 3017->3058, OAuthModal.tsx 969->989, cliRuntime.ts 1090->1100, webProvidersA.ts 805->809. Test frozen raised: deepseek-web.test.ts 1081->1092. Real sizes (check-file-size.mjs reported). These stay frozen (cannot grow further); structural shrink tracked under decomposition roadmap #3501; the release captain's rebaseline-at-release supersedes this note. Bundled with the #5695 quick-start test regex fix (multi-line <Link> tolerance) in the same release-green PR.",
   "_rebaseline_2026_07_02_5798_release_green": "Release-green unblock #5798 / PR #5896 (2026-07-02): the quality.yml fast-gates job was base-red for EVERY PR->release (whole queue failing), from growth inherited via already-merged PRs — no offending PR branch left to fix. Prod frozen raised: AddApiKeyModal.tsx 869->905, providerPageHelpers.ts 996->1021, RequestLoggerV2.tsx 1316->1553, src/sse/services/auth.ts 2403->2405, antigravity.ts 1806->1813, base.ts 1502->1536 (1533 inherited + 3 lines from this PR's own typecheck:core fix in resolveBaseUrl), advancedTools.ts 1118->1120, accountFallback.ts 1783->1790, openai-to-kiro.ts 842->853, openai-responses.ts 1035->1092, stream.ts 2710->2727; new-above-cap frozen: webProvidersA.ts 805, tokenHealthCheck.ts 830. Test frozen raised: cc-compatible-provider 1179->1217, translator-openai-to-kiro 999->1088, web-cookie-providers-new 827->845; new-above-cap: response-sanitizer.test.ts 906. These files remain frozen (cannot grow further); the release captain's rebaseline-at-release supersedes this note.",
@@ -193,7 +194,7 @@
     "open-sse/services/usage.ts": 3454,
     "open-sse/translator/request/openai-to-gemini.ts": 906,
     "open-sse/translator/request/openai-to-kiro.ts": 912,
-    "open-sse/translator/response/openai-responses.ts": 1092,
+    "open-sse/translator/response/openai-responses.ts": 1112,
     "open-sse/utils/cursorAgentProtobuf.ts": 1521,
     "open-sse/utils/stream.ts": 2796,
     "src/app/(dashboard)/dashboard/HomePageClient.tsx": 1385,
